New every Nth step Matrix
Creates a new array that stores every 20th step of the bigger array.
Right now the x and the y alternate one off but I’m not sure why at the
moment. Also a new file reader was made and the second matrix can now
print into a cvs file in a different folder than the first one.